iota_graphql_rpc::test_infra

Module cluster

Structs

Cluster

ExecutorCluster

SnapshotLagConfig

Constants

DEFAULT_INTERNAL_DATA_SOURCE_PORT

Functions

serve_executor

Takes in a simulated instantiation of an IOTA blockchain and builds a cluster around it.

start_cluster

Starts a validator, fullnode, indexer, and graphql service for testing.

start_graphql_server_with_fn_rpc

wait_for_graphql_checkpoint_pruned

Ping the GraphQL server for a checkpoint until an empty response is returned, indicating that the checkpoint has been pruned.