iota_graphql_rpc::test_infra::cluster

Function wait_for_graphql_checkpoint_pruned

pub async fn wait_for_graphql_checkpoint_pruned(
    client: &SimpleClient,
    checkpoint: u64,
    base_timeout: Duration,
)

Ping the GraphQL server for a checkpoint until an empty response is returned, indicating that the checkpoint has been pruned.