Module iota_graphql_rpc::test_infra::cluster

Structs

Cluster

ExecutorCluster

SnapshotLagConfig

Constants

DEFAULT_INTERNAL_DATA_SOURCE_PORT

Functions

serve_executor

Takes in a simulated instantiation of a Iota blockchain and builds a cluster around it. This cluster is typically used in e2e tests to emulate and test behaviors.

start_cluster

Starts a validator, fullnode, indexer, and graphql service for testing.

start_graphql_server

start_graphql_server_with_fn_rpc